Trait sc_client::BlockchainEvents

pub trait BlockchainEvents<Block> where
    Block: Block, {
    fn import_notification_stream(
        &self
    ) -> UnboundedReceiver<BlockImportNotification<Block>>;

    fn finality_notification_stream(
        &self
    ) -> UnboundedReceiver<FinalityNotification<Block>>;

    fn storage_changes_notification_stream(
        &self, 
        filter_keys: Option<&[StorageKey]>, 
        child_filter_keys: Option<&[(StorageKey, Option<Vec<StorageKey>>)]>
    ) -> Result<UnboundedReceiver<(<Block as Block>::Hash, StorageChangeSet)>, Error>;
}

A source of blockchain events.

Required methods

fn import_notification_stream(
&self
) -> UnboundedReceiver<BlockImportNotification<Block>>

Get block import event stream. Not guaranteed to be fired for every imported block.

fn finality_notification_stream(
&self
) -> UnboundedReceiver<FinalityNotification<Block>>

Get a stream of finality notifications. Not guaranteed to be fired for every finalized block.

fn storage_changes_notification_stream(
&self,
filter_keys: Option<&[StorageKey]>,
child_filter_keys: Option<&[(StorageKey, Option<Vec<StorageKey>>)]>
) -> Result<UnboundedReceiver<(<Block as Block>::Hash, StorageChangeSet)>, Error>

Get storage changes event stream.

Passing None as filter_keys subscribes to all storage changes.

Loading content...